improve calculation accuracy
Allow using float point arguments to avoid truncation.

Test: Add (4.9, 4.9) should return 10.

fix truncation error

To convert float to integer the truncation is performed, but the
rounding is expected.

Test: Add (4.9, 4.9) should return 10.

formatting: use tabs instead of spaces

+++ calculator.h
@@ -3,9 +3,9 @@
 
 class Calculator
 {
-    public:
-        int Add (int, int);
-        int Sub (int, int);
+	public:
+		int Add (double, double);
+		int Sub (double, double);
 };
 
 #endif//CALCULATOR_H
